refactor (GetDetailedProfileResponse): tag 정보 nested class 로 변경
